chore: refactor move msg & state structs to package
