+++ contracts/cw-ics20-latest/src/state.rs
@@ -1,12 +1,12 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{Addr, IbcEndpoint, StdResult, Storage, Uint128};
+use cosmwasm_std::{Addr, StdResult, Storage, Uint128};
 use cw20_ics20_msg::converter::ConverterController;
+use cw20_ics20_msg::state::{
+    AllowInfo, ChannelInfo, ConvertReplyArgs, MappingMetadata, Ratio, ReplyArgs,
+};
 use cw_controllers::Admin;
 use cw_storage_plus::{Index, IndexList, IndexedMap, Item, Map, MultiIndex};
-use oraiswap::{
-    asset::{Asset, AssetInfo},
-    router::RouterController,
-};
+use oraiswap::router::RouterController;
 
 use crate::ContractError;
 
@@ -95,63 +95,6 @@ pub struct Config {
     pub token_fee_receiver: Addr,
     pub relayer_fee_receiver: Addr,
     pub converter_contract: ConverterController,
-}
-
-#[cw_serde]
-pub struct ChannelInfo {
-    /// id of this channel
-    pub id: String,
-    /// the remote channel/port we connect to
-    pub counterparty_endpoint: IbcEndpoint,
-    /// the connection this exists on (you can use to query client/consensus info)
-    pub connection_id: String,
-}
-
-#[cw_serde]
-pub struct AllowInfo {
-    pub gas_limit: Option<u64>,
-}
-
-#[cw_serde]
-pub struct TokenFee {
-    pub token_denom: String,
-    pub ratio: Ratio,
-}
-
-#[cw_serde]
-pub struct RelayerFee {
-    pub prefix: String,
-    pub fee: Uint128,
-}
-
-#[cw_serde]
-pub struct Ratio {
-    pub nominator: u64,
-    pub denominator: u64,
-}
-
-#[cw_serde]
-pub struct MappingMetadata {
-    /// asset info on local chain. Can be either cw20 or native
-    pub asset_info: AssetInfo,
-    pub remote_decimals: u8,
-    pub asset_info_decimals: u8,
-    #[serde(default)]
-    pub is_mint_burn: bool,
-}
-
-#[cw_serde]
-pub struct ReplyArgs {
-    pub channel: String,
-    pub local_receiver: String,
-    pub denom: String,
-    pub amount: Uint128,
-}
-
-#[cw_serde]
-pub struct ConvertReplyArgs {
-    pub local_receiver: String,
-    pub asset: Asset,
 }
 
 pub fn increase_channel_balance(
