+++ packages/cw20-ics20-msg/src/lib.rs
@@ -6,4 +6,6 @@ pub mod amount;
 pub mod converter;
 pub mod helper;
 pub mod ibc_hooks;
+pub mod msg;
 pub mod receiver;
+pub mod state;
